UI刷新と責務分割のリファクタリング
