+++ README.md
@@ -1,11 +1,20 @@
 # rough_to_illustration
 
-生成AIでラフ絵からイラストを作成するFlaskアプリケーション。ログイン後にラフ絵とプロンプトを送信し、Nano Banana(Gemini API)で生成したイラストをプレビュー・ダウンロードできます。
+生成AIでラフ絵からイラストを作成するFlaskアプリケーション。ログイン後にラフ絵とプロンプトを送信し、Nano Banana(Gemini API)で生成したイラストをプレビュー・ダウンロードできます。アップロード時のプレビュー、文字数カウンタ、送信中のローディング表示などリッチなUIを備えています。
 
 ## 主な機能
 - Flask + SQLAlchemy + Flask-Login によるログイン必須のWeb UI
 - ラフ絵アップロードと色指定・ポーズ指示、任意のアスペクト比/解像度入力
+- アップロードプレビュー、文字数カウンタ、送信中のローディング表示を備えたリッチUI
 - Gemini API を呼び出して生成した画像のプレビューとダウンロード
+
+## ディレクトリ構成
+- `app.py`: Flaskアプリのエントリポイント。Blueprint登録のみを担当。
+- `config.py` / `extensions.py` / `models.py`: 設定値、拡張機能のインスタンス、ユーザーモデル。
+- `services/`: プロンプト生成や画像生成処理のヘルパー。
+- `views/`: 認証とメイン画面のBlueprint。
+- `templates/`: リッチなUIを定義したJinja2テンプレート。
+- `old/streamlit/`: 旧Streamlitモックアップのコードを保管（現行アプリでは未使用）。
 
 ## 事前準備
 1. 依存関係のインストール
@@ -33,6 +42,11 @@
 flask --app app.py run  # または python app.py
 ```
 ブラウザで `http://localhost:5000/` にアクセスし、登録済みユーザーでログインしてください。ログインしないと生成フォームは利用できません。
+
+### UIのヒント
+- ラフ絵をアップロードするとプレビューが表示されます。不要になった場合は「画像をクリア」でリセットできます。
+- テキスト入力には文字数カウンタがあり、推奨の文字数ガイドを併記しています。
+- 送信中はボタンが自動で無効化され、スピナーが表示されます。
 
 ## 使い方
 1. ログイン後、ラフ絵(PNG/JPEG)を選択します。
