+++ templates/base.html
@@ -4,38 +4,138 @@
     <meta charset="utf-8">
     <meta name="viewport" content="width=device-width, initial-scale=1">
     <title>{{ title or 'ラフ絵からイラスト生成' }}</title>
+    <link rel="preconnect" href="https://fonts.googleapis.com">
+    <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin>
+    <link href="https://fonts.googleapis.com/css2?family=Inter:wght@400;600;700&family=Noto+Sans+JP:wght@400;500;700&display=swap" rel="stylesheet">
     <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.3/dist/css/bootstrap.min.css">
+    <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bootstrap-icons@1.11.3/font/bootstrap-icons.css">
+    <style>
+      :root {
+        --primary: #7c4dff;
+        --primary-strong: #5a2ded;
+        --accent: #1fd1f9;
+        --bg-dark: #0c0f1a;
+        --glass: rgba(255, 255, 255, 0.08);
+      }
+      * { font-family: 'Inter', 'Noto Sans JP', system-ui, sans-serif; }
+      body {
+        background: radial-gradient(circle at 20% 20%, rgba(124, 77, 255, 0.18), transparent 35%),
+                    radial-gradient(circle at 80% 0%, rgba(31, 209, 249, 0.24), transparent 25%),
+                    linear-gradient(135deg, #0c1220 0%, #0b0f19 100%);
+        color: #f7f9ff;
+        min-height: 100vh;
+      }
+      .navbar {
+        background: rgba(12, 15, 26, 0.85);
+        backdrop-filter: blur(10px);
+        border-bottom: 1px solid rgba(255, 255, 255, 0.08);
+      }
+      .navbar-brand {
+        font-weight: 700;
+        letter-spacing: 0.02em;
+      }
+      .glass-card {
+        background: var(--glass);
+        border: 1px solid rgba(255, 255, 255, 0.08);
+        border-radius: 18px;
+        box-shadow: 0 15px 60px rgba(0, 0, 0, 0.35);
+        backdrop-filter: blur(14px);
+      }
+      .glass-card .card-body { color: #e8ecff; }
+      .subtle-text { color: #c7cce5; }
+      .badge-soft {
+        background: rgba(124, 77, 255, 0.12);
+        color: #d6c4ff;
+        border: 1px solid rgba(124, 77, 255, 0.35);
+      }
+      .hero-title {
+        font-size: clamp(1.7rem, 2vw, 2rem);
+        font-weight: 700;
+      }
+      .btn-primary {
+        background: linear-gradient(135deg, var(--primary), var(--accent));
+        border: none;
+        box-shadow: 0 12px 30px rgba(124, 77, 255, 0.35);
+      }
+      .btn-primary:hover { background: linear-gradient(135deg, var(--primary-strong), var(--accent)); }
+      .btn-outline-light:hover { color: #0c0f1a; background-color: #f7f9ff; }
+      .form-control, .form-select, textarea {
+        background: rgba(255, 255, 255, 0.08);
+        border: 1px solid rgba(255, 255, 255, 0.1);
+        color: #f7f9ff;
+      }
+      .form-control:focus, .form-select:focus, textarea:focus {
+        border-color: var(--accent);
+        box-shadow: 0 0 0 0.1rem rgba(31, 209, 249, 0.3);
+        background: rgba(255, 255, 255, 0.12);
+      }
+      .flash-stack {
+        position: sticky;
+        top: 12px;
+        z-index: 10;
+      }
+      .flash-card {
+        border-radius: 14px;
+        border: 1px solid rgba(255, 255, 255, 0.15);
+        background: rgba(12, 15, 26, 0.75);
+        color: #f7f9ff;
+        box-shadow: 0 10px 40px rgba(0, 0, 0, 0.25);
+      }
+      .flash-card .btn-close { filter: invert(1); }
+      .footer-note {
+        color: #9aa2c5;
+        font-size: 0.9rem;
+      }
+    </style>
   </head>
-  <body class="bg-light">
-    <nav class="navbar navbar-expand-lg navbar-dark bg-dark mb-4">
-      <div class="container">
-        <a class="navbar-brand" href="{{ url_for('index') }}">ラフ絵 to イラスト</a>
-        <div class="d-flex">
+  <body>
+    <nav class="navbar navbar-expand-lg navbar-dark mb-4">
+      <div class="container py-2">
+        <a class="navbar-brand" href="{{ url_for('main.index') }}">ラフ絵 to イラスト</a>
+        <div class="d-flex align-items-center gap-2 ms-auto">
           {% if current_user.is_authenticated %}
-          <span class="navbar-text text-white me-3">{{ current_user.username }} さん</span>
-          <a class="btn btn-outline-light" href="{{ url_for('logout') }}">ログアウト</a>
+          <span class="badge rounded-pill bg-light text-dark me-1"><i class="bi bi-person"></i> {{ current_user.username }} さん</span>
+          <a class="btn btn-outline-light" href="{{ url_for('auth.logout') }}">ログアウト</a>
           {% else %}
-          <a class="btn btn-outline-light me-2" href="{{ url_for('login') }}">ログイン</a>
-          <a class="btn btn-primary" href="{{ url_for('signup') }}">新規登録</a>
+          <a class="btn btn-outline-light me-2" href="{{ url_for('auth.login') }}">ログイン</a>
+          <a class="btn btn-primary" href="{{ url_for('auth.signup') }}">新規登録</a>
           {% endif %}
         </div>
       </div>
     </nav>
 
-    <div class="container">
-      {% with messages = get_flashed_messages(with_categories=true) %}
-        {% if messages %}
-        <div class="mb-3">
-          {% for category, message in messages %}
-          <div class="alert alert-{{ 'danger' if category == 'error' else category }}" role="alert">
-            {{ message }}
+    <div class="container pb-5">
+      <div class="flash-stack">
+        {% with messages = get_flashed_messages(with_categories=true) %}
+          {% if messages %}
+          <div class="d-flex flex-column gap-2">
+            {% for category, message in messages %}
+            <div class="alert flash-card alert-{{ 'danger' if category == 'error' else category }} alert-dismissible fade show" role="alert">
+              <i class="bi bi-info-circle me-2"></i>{{ message }}
+              <button type="button" class="btn-close" data-bs-dismiss="alert" aria-label="Close"></button>
+            </div>
+            {% endfor %}
           </div>
-          {% endfor %}
-        </div>
-        {% endif %}
-      {% endwith %}
+          {% endif %}
+        {% endwith %}
+      </div>
 
       {% block content %}{% endblock %}
+
+      <div class="mt-5 text-center footer-note">データはセッションにのみ保持され、ログアウトでリセットされます。</div>
     </div>
+
+    <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.3/dist/js/bootstrap.bundle.min.js"></script>
+    <script>
+      const autoDismiss = () => {
+        document.querySelectorAll('.alert.flash-card').forEach((alertEl) => {
+          setTimeout(() => {
+            const alertInstance = bootstrap.Alert.getOrCreateInstance(alertEl);
+            alertInstance.close();
+          }, 5200);
+        });
+      };
+      document.addEventListener('DOMContentLoaded', autoDismiss);
+    </script>
   </body>
 </html>
