+++ templates/login.html
@@ -1,21 +1,36 @@
 {% extends 'base.html' %}
 {% block content %}
 <div class="row justify-content-center">
-  <div class="col-md-6">
-    <div class="card shadow-sm">
-      <div class="card-body">
-        <h2 class="card-title mb-4">ログイン</h2>
-        <form method="post">
-          <div class="mb-3">
+  <div class="col-lg-6 col-xl-5">
+    <div class="card glass-card shadow-lg">
+      <div class="card-body p-4 p-lg-5">
+        <div class="d-flex justify-content-between align-items-center mb-3">
+          <div>
+            <div class="badge badge-soft mb-2">サインイン</div>
+            <h2 class="card-title mb-1">ようこそ、おかえりなさい</h2>
+            <p class="subtle-text mb-0">登録済みのユーザー名とパスワードでログインしてください。</p>
+          </div>
+          <i class="bi bi-door-open fs-2 text-info"></i>
+        </div>
+        <form method="post" class="d-flex flex-column gap-3">
+          <div>
             <label class="form-label" for="username">ユーザー名</label>
-            <input class="form-control" id="username" name="username" required>
+            <div class="input-group">
+              <span class="input-group-text bg-transparent text-white-50"><i class="bi bi-person"></i></span>
+              <input class="form-control" id="username" name="username" required placeholder="demo_user">
+            </div>
           </div>
-          <div class="mb-3">
+          <div>
             <label class="form-label" for="password">パスワード</label>
-            <input class="form-control" id="password" type="password" name="password" required>
+            <div class="input-group">
+              <span class="input-group-text bg-transparent text-white-50"><i class="bi bi-shield-lock"></i></span>
+              <input class="form-control" id="password" type="password" name="password" required placeholder="••••••••">
+            </div>
           </div>
-          <button class="btn btn-primary" type="submit">ログイン</button>
-          <a class="btn btn-link" href="{{ url_for('signup') }}">新規登録はこちら</a>
+          <div class="d-grid gap-2 d-md-flex justify-content-md-between align-items-center">
+            <button class="btn btn-primary px-4" type="submit">ログイン</button>
+            <a class="text-decoration-none subtle-text" href="{{ url_for('auth.signup') }}">新規登録はこちら</a>
+          </div>
         </form>
       </div>
     </div>
